requirements.txt
streamlit
opencv-python
transformers
gtts
torch
Pillow
